requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pytz==2021.1
sqlparse==0.4.1
PyJWT==2.1.0
PyYAML==5.4.1
uritemplate==3.0.1
urllib3==1.26.6
idna==3.2
charset-normalizer==2.0.1
certifi==2021.5.30
requests==2.26.0
six==1.16.0
text-unidecode==1.3
python-dateutil==2.8.2
Faker==8.10.1
